Crate linfa_lars

Least angle regression a.k.a. LAR

This struct contains the parameters of a fitted LARS model. This includes the seperating hyperplane, (optionally) intercept, alphas (Maximum of covariances (in absolute value) at each iteration), Indices of active variables at the end of the path,

LARS is similar to forward stepwise regression. At each step, it finds the feature most correlated with the target. When there are multiple features having equal correlation, instead of continuing along the same feature, it proceeds in a direction equiangular between the features.

References

“Least Angle Regression”, Efron et al.

Wikipedia entry on the Least-angle regression

Scikit-Learn User Guide

Examples

use linfa::prelude::*;
use linfa_lars::Lars;

// Load the Diabetes dataset
let dataset = linfa_datasets::diabetes();

// Create a LARS model with intercept fitting enabled
let lars = Lars::params().fit_intercept(true).fit(&dataset).unwrap();

// Generate predictions using the trained model and compute the R² score of the predictions
let pred = lars.predict(&dataset);
let r2 = pred.r2(&dataset).unwrap();
println!("r2 from prediction: {}", r2);

Structs

Lars

LarsParams

LarsValidParams

Enums

LarsError